Fix get metals request
A problem with empty return values was not because of conman library.
The poin was in H2 recursive query limitations and SQL parameters
binding by HugSQL. The problem is solved by using user defined vars in
H2 database. This fix add update-id-param query, update get-metals-by-id
query and remove unused get-metals.